+++ apps/api/src/app/portfolio/portfolio.service.ts
@@ -24,7 +24,7 @@ import {
   MAX_CHART_ITEMS,
   UNKNOWN_KEY
 } from '@ghostfolio/common/config';
-import { DATE_FORMAT, parseDate } from '@ghostfolio/common/helper';
+import { DATE_FORMAT, getSum, parseDate } from '@ghostfolio/common/helper';
 import {
   Accounts,
   EnhancedSymbolProfile,
@@ -678,6 +678,8 @@ export class PortfolioService {
       return {
         tags,
         averagePrice: undefined,
+        dividendInBaseCurrency: undefined,
+        feeInBaseCurrency: undefined,
         firstBuyDate: undefined,
         grossPerformance: undefined,
         grossPerformancePercent: undefined,
@@ -744,11 +746,22 @@ export class PortfolioService {
         averagePrice,
         currency,
         dataSource,
+        fee,
         firstBuyDate,
         marketPrice,
         quantity,
         transactionCount
       } = position;
+
+      const dividendInBaseCurrency = getSum(
+        orders
+          .filter(({ type }) => {
+            return type === 'DIVIDEND';
+          })
+          .map(({ valueInBaseCurrency }) => {
+            return new Big(valueInBaseCurrency);
+          })
+      );
 
       // Convert investment, gross and net performance to currency of user
       const investment = this.exchangeRateDataService.toCurrency(
@@ -836,6 +849,12 @@ export class PortfolioService {
         tags,
         transactionCount,
         averagePrice: averagePrice.toNumber(),
+        dividendInBaseCurrency: dividendInBaseCurrency.toNumber(),
+        feeInBaseCurrency: this.exchangeRateDataService.toCurrency(
+          fee.toNumber(),
+          SymbolProfile.currency,
+          userCurrency
+        ),
         grossPerformancePercent:
           position.grossPerformancePercentage?.toNumber(),
         historicalData: historicalDataArray,
@@ -892,6 +911,8 @@ export class PortfolioService {
         SymbolProfile,
         tags,
         averagePrice: 0,
+        dividendInBaseCurrency: 0,
+        feeInBaseCurrency: 0,
         firstBuyDate: undefined,
         grossPerformance: undefined,
         grossPerformancePercent: undefined,